Add entity tracker and item pickup/despawn
Introduce an EntityTracker to track players, mobs, items and orbs and expose track/untrack/update/is_visible utilities. Extend ServerContext with entity_tracker, item_spawn_times, item_positions, despawn_tx and pickup_tx, and spawn a background task to despawn items after world.item_despawn_seconds. Add inventory support (Inventory, Slot, SetSlot) and client/server handling for item drops, stacking, pickup (CollectItem + SetSlot + DestroyEntities), and held-slot tracking; spawn/tracking information is emitted to other players only when visible. Update player join flow to accept a UUID and register tracked player entities; untrack on disconnect. Add TrackingConfig and world.item_despawn_seconds config defaults and sample values. Misc: small protocol packet additions (CollectItem/SetSlot), ChatBuilder string formatting simplifications, clamp held slot, deny sprint when low on food, and a chunk coordinate fix in chunk builder.

+++ src/world/chunk.rs
@@ -99,7 +99,7 @@ async fn build_chunk_data_18(
                 for x in 0..16usize {
                     let wx = chunk_x * 16 + x as i32;
                     let wz = chunk_z * 16 + z as i32;
-                    let state = get_block_state(&snapshot, wx, y as u8, wz as i32);
+                    let state = get_block_state(&snapshot, wx, y as u8, wz);
                     data.extend_from_slice(&state.to_le_bytes());
                 }
             }
